+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# NEXT
+# 3.0.0
 
 ## New Features